Move `pkg/driver` to `pkg/icingadb/driver.go`
