+++ config/database.go
@@ -6,7 +6,6 @@ import (
 	"database/sql/driver"
 	"fmt"
 	"github.com/go-sql-driver/mysql"
-	icingadbDriver "github.com/icinga/icingadb/pkg/driver"
 	"github.com/icinga/icingadb/pkg/icingadb"
 	"github.com/icinga/icingadb/pkg/logging"
 	"github.com/icinga/icingadb/pkg/utils"
@@ -40,7 +39,7 @@ type Database struct {
 // calls sqlx.Open, but returns *icingadb.DB.
 func (d *Database) Open(logger *logging.Logger) (*icingadb.DB, error) {
 	registerDriverOnce.Do(func() {
-		icingadbDriver.Register(logger)
+		icingadb.Register(logger)
 	})
 
 	var db *sqlx.DB
@@ -89,7 +88,7 @@ func (d *Database) Open(logger *logging.Logger) (*icingadb.DB, error) {
 			return setGaleraOpts(ctx, conn, wsrepSyncWait)
 		}
 
-		db = sqlx.NewDb(sql.OpenDB(icingadbDriver.NewConnector(c, logger, setWsrepSyncWait)), icingadbDriver.MySQL)
+		db = sqlx.NewDb(sql.OpenDB(icingadb.NewConnector(c, logger, setWsrepSyncWait)), icingadb.MySQL)
 	case "pgsql":
 		uri := &url.URL{
 			Scheme: "postgres",
@@ -143,7 +142,7 @@ func (d *Database) Open(logger *logging.Logger) (*icingadb.DB, error) {
 			return nil, errors.Wrap(err, "can't open pgsql database")
 		}
 
-		db = sqlx.NewDb(sql.OpenDB(icingadbDriver.NewConnector(connector, logger, nil)), icingadbDriver.PostgreSQL)
+		db = sqlx.NewDb(sql.OpenDB(icingadb.NewConnector(connector, logger, nil)), icingadb.PostgreSQL)
 	default:
 		return nil, unknownDbType(d.Type)
 	}
